requirements.txt
cozmo>=0.9.0
cozmoclad>=1.1.0
Pillow>=3.4.2